Revert "Validation Project v2 (#444)"
This reverts commit b2e1e553e419dc7bd2212e347a9acedaa2464595.
