--- src/Ryujinx.BuildValidationTasks/ValidationTask.cs
@@ -1,7 +1,0 @@
-namespace Ryujinx.BuildValidationTasks
-{
-    public interface ValidationTask
-    {
-        public bool Execute(string projectPath, bool isGitRunner);
-    }
-}
